requirements.txt
# Copyright (C) 2020-2022 Mitsubishi Electric Research Laboratories (MERL)
#
# SPDX-License-Identifier: AGPL-3.0-or-later

audioread==2.1.8
cffi==1.13.2
cycler==0.10.0
decorator==4.4.1
future==0.18.2
gputil==1.4.0
h5py==2.10.0
joblib==0.14.1
kiwisolver==1.1.0
librosa==0.7.1
llvmlite==0.31.0
loguru==0.4.0
matplotlib==3.1.2
numba==0.47.0
numpy==1.18.1
pandas==0.25.3
pillow==7.0.0
pycparser==2.19
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
pyyaml==5.3
resampy==0.2.2
scikit-learn==0.20.4
scipy==1.4.1
six==1.13.0
soundfile==0.10.3.post1
torch==1.3.1
torchvision==0.4.2
tensorboard==2.2.1
tqdm
